requirements.txt
google-generativeai>=0.8.5
browser-use>=0.1.41
playwright>=1.51.0
open-interpreter>=0.4.3
vosk>=0.3.45
SpeechRecognition>=3.14.2
pyttsx3>=2.98
python-dotenv>=1.1.0
pydantic>=2.11.3
rich>=14.0.0
typer>=0.15.2
jsonschema>=4.23.0
uv>=0.6.16 